Bump \p{nv=} precision from 2 to 3
This closes #19603

Unicode has various characters whose numeric value is rational
non-integer.   These can be specified in \p{nv=...} constructs by either
the rational form or by an expression that it evaluates to.  The number
of significant digits that must match are kept to a minimum to allow for
variances in different platforms floating point lengths and rounding
decisions.  Previously that number was 2 digits; but that is no longer
always sufficient for all platforms.  This commit changes it to 3.